+++ test/VaultERC20.t.sol
@@ -20,8 +20,8 @@ contract VaultERC20Test is Test {
         xyk = new XtYK();
         token0 = new MockERC20("", "", 18);
         token1 = new MockERC20("", "", 18);
-        token0.mint(address(0xBEEF), type(uint96).max);
-        token1.mint(address(0xBEEF), type(uint96).max);
+        token0.mint(address(0xBEEF), type(uint128).max);
+        token1.mint(address(0xBEEF), type(uint128).max);
         vm.startPrank(address(0xBEEF));
         token0.approve(address(vault), type(uint256).max);
         token1.approve(address(vault), type(uint256).max);
